# Add heuristics