1. Overall tactic execution: 61% faster (602ms → 234ms)
  2. Simp performance: 70% faster (146ms → 44ms)
  3. Norm_num performance: 39% faster (20.8ms → 12.7ms)